# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API Documentation",
        "description": "L5 Swagger API Documentation for Laravel 10",
        "version": "1.0.0"
    },
    "paths": {
        "/api/profiles/{profile}/comments": {
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Add a comment to a profile",
                "description": "Allows an authenticated administrator to post a comment on a profile",
                "operationId": "storeComment",
                "parameters": [
                    {
                        "name": "profile",
                        "in": "path",
                        "description": "ID of the profile",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content",
                                    "profile_id"
                                ],
                                "properties": {
                                    "content": {
                                        "type": "string",
                                        "example": "This is a comment"
                                    },
                                    "profile_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Comment added successfully"
                    },
                    "400": {
                        "description": "You have already commented on this profile"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}